made a separate compononet for todos and changed the imports accordingly
