+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './App.css';
 import Header from './components/layout/Header';
-import Todo from './components/Todos';
+import Todo from './components/todo/Todos';
 
 class App extends Component {
